window: Add login page to main stack
Previously, a login page was added as the placeholder of timeline
list. But every time you open the program, the login page shows up,
no matter whether you have logged in or not. You have to open the login
dialog and close it to fetch a list of Weibo posts.

Fix that by moving the login page inside the main stack and checking
access token for whether to show login page.