image-button: Improve image scaling for thumbnails
Previously images are scaled down by simply calling
gdk_pixbuf_scale_simple(), which doesn’t handle the aspect ratio very
well. This leads to unnatural look of the thumbnails.

The thumbnail is shown in 150px by 150px, which is square. To fix the
issue mentioned above, we crop the central square part of the image and
them scale it down to 150px by 150px. Images smaller than 150px by 150px
are still simply scaled, which is not an issue.