Rename the project to Weibird
This is a GTK+ client for viewing Weibo, which is a platform where you
can tweet(post) interesting things. "tweet" originally means "the chirp of
a small or young bird". That’s where bird comes from.

So Weibird is a program that you can tweet(post) in Weibo (weibo.com).