media-image: Remove WbMediaImage class
The image process was moved back to WbImageButton. We don’t really use
WbMediaImage while showing larger images in media dialog. We get the
pixbuf owned by the image button, crop it to the size we want and display
it. WbMediaImage requires an url and fetches the remote image. So using
that will cause fetching images which we already own in the image button.

Revert 8af4fae9.