media-dialog: Resize dialog while switching images
Previously, if you’re viewing an image with smaller width and switch to
another larger image, the media dialog won’t resize to properly show
the image. You have to scroll to see the whole image.

Fix that by resizing the dialog while switching images.

This is not a ultimate fix by any mean. Ideally, the size of media dialog
should be the same as the image shown in it.