util: Fix date showing up for posts from today
Previously, when it’s one of the days from the 1st to 9th of a month, the
day of the month is displayed as well, which is not necessary. It’s the
same day as it currently is.

Fix that by prefixing the day string with “0” when it’s smaller than 10.
This matches the time format used by Weibo API time, like
Wed Mar 06 14:00:58 +0800 2019.