image-button: Scale middle quality image as thumbnail
Previously, we load the original(full quality) image and scale it down
as thumbnail, which consumes lots of memory. If you don’t click the image
and view the original image, it doesn’t make any sense to do so.

Thumbnail size provided by Weibo API is usually too small for our
thumbnail use. So load the middle quality instead of original quality
image and scale it down to save memory consumption.