media-dialog: Only load images when they are shown
Previously, all images in the tweet are already loaded in max quality
before we open them in the dialog. It’s memory-consuming, and useless if
we don’t even click any image to view them in the dialog.

To fix that, we pass a list of uris for images in the tweet, and only
load the image that’s shown currently. The loaded image will be freed if
you choose to view the previous/next image if there is one.