main-widget: Add a loading indicator on application startup
If you have a slow network, loading WbTimelineList can be time
consuming. In this case, showing the application window takes quite a
long time.

To fix that, present the window even though timeline list is not fully
loaded. Beside, show a "loading" label to indicate that the application
is working hard to load something.